requirements.txt
pyside6
selenium
webdriver-manager
python-dotenv